backend/requirements.txt
mediapipe
opencv-python
paho-mqtt
pillow
numpy
